incremented version and moved some code
