+++ internal/score/quartettotals.go
